backend/requirements.txt
argh==0.26.2
argon2-cffi==19.1.0
asn1crypto==0.24.0
attrs==19.1.0
autobahn==19.7.1
Automat==0.7.0
bcrypt==3.1.7
bitstring==3.1.5
cbor==1.0.0
cbor2==4.1.2
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
constantly==15.1.0
crossbar==19.6.2
cryptography==2.7
flatbuffers==1.11
h2==3.1.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==19.0.0
idna==2.5
incremental==17.5.0
Jinja2==2.10.1
lmdb==0.95
lz4==2.1.10
MarkupSafe==1.1.1
mistune==0.8.4
msgpack==0.6.1
netaddr==0.7.19
numpy==1.15.4
paho-mqtt==1.4.0
passlib==1.7.1
pathtools==0.1.2
priority==1.3.0
psutil==5.6.3
py-ubjson==0.14.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycparser==2.19
Pygments==2.4.2
PyHamcrest==1.9.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
PyQRCode==1.2.1
python-snappy==0.5.4
PyTrie==0.3.1
PyYAML==5.1.1
requests==2.22.0
sdnotify==0.3.2
service-identity==18.1.0
setproctitle==1.1.10
six==1.12.0
sortedcontainers==2.1.0
treq==18.6.0
Twisted==19.2.1
txaio==18.8.1
txtorcon==19.0.0
u-msgpack-python==2.5.1
ujson==1.35
urllib3==1.24.3
watchdog==0.9.0
zlmdb==19.6.2
zope.interface==4.6.0
